♻️ Move track_run_inputs test in its own file
